requirements.txt
pyyaml
httpx
numpy
colorama
pyfiglet